Polishes code

+++ src/pkgmngr/manager.py
@@ -5,60 +5,57 @@ try:
     from pkgmngr import capsule as caps
     from pkgmnger import registry as reg
     from pkgmnger import graph
+    from pkgmnger import apparatus as apt
 except:
     import capsule as caps
     import registry as reg
+    import apparatus as apt
     import graph
 
 class legoHDL:
     def __init__(self):
-
         self.capsulePKG = None
         self.capsuleCWD = None
-        #defines path to working dir of 'legoHDL' tool
-        self.pkgmngPath = os.path.realpath(__file__)
-        self.pkgmngPath = self.pkgmngPath[:self.pkgmngPath.rfind('/')+1]
-
-        self.settings = dict()
-        with open(self.pkgmngPath+"settings.yml", "r") as file:
-            self.settings = yaml.load(file, Loader=yaml.FullLoader)
         
         # !!! UNCOMMENT LINE BELOW TO DISABLE REMOTE !!!
         #self.settings['remote'] = None #testing allowing option to not connect to a remote!
         #defines path to dir of remote code base
-        self.remote = self.settings['remote']
+        self.db = reg.Registry(apt.Apparatus.SETTINGS['remote'])
 
-        #set class capsule variable for user settings
-        caps.Capsule.settings = self.settings
-        caps.Capsule.pkgmngPath = self.pkgmngPath
-        
-        self.db = reg.Registry(self.remote)
-        
-        #defines path to dir of local code base
-        self.local = os.path.expanduser(self.settings['local'])+"/"
-        self.hidden = os.path.expanduser("~/.legohdl/") #path to registry and cache
         #defines how to open workspaces
-        self.textEditor = self.settings['editor']
+        self.textEditor = apt.Apparatus.SETTINGS['editor']
 
-        self.db.findProjectsLocal(self.local)
+        self.db.getProjectsLocal()
+        #exit()
+
+        self.db.findProjectsLocal(apt.Apparatus.SETTINGS['local'])
         if(caps.Capsule.linkedRemote()): #fetch remote servers
             self.db.fetch()
         self.db.sync()
-        
-        os.environ['VHDL_LS_CONFIG'] = self.hidden+"map.toml" #directly works with VHDL_LS
+        #directly works with VHDL_LS
+        os.environ['VHDL_LS_CONFIG'] = apt.Apparatus.HIDDEN+"map.toml" 
 
         self.parse()
         pass
 
-    def genPKG(self, cap):
-        lib_path = self.hidden+"lib/"+cap.getLib()+"/"
+    def genPKG(self, title):
+        cap = None
+        if(self.db.prjExists(title, "cache")):
+            cache = self.db.getProjectsCache()
+            l,n = caps.Capsule.siftLibName(title)
+            cap = caps.Capsule(title=title,path=cache[l][n])
+        else:
+            print("Error- the project is not located in the cache")
+            return
+
+        lib_path = apt.Apparatus.HIDDEN+"lib/"+cap.getLib()+"/"
         os.makedirs(lib_path, exist_ok=True)
-        tmp_pkg = open(self.pkgmngPath+"/template_pkg.vhd", 'r')
+        tmp_pkg = open(apt.Apparatus.PKGMNG_PATH+"/template_pkg.vhd", 'r')
         vhd_pkg = open(lib_path+cap.getName()+"_pkg.vhd", 'w')
 
         #need to look at toplevel VHD file to transfer correct library uses
         #search through all library uses and see what are chained dependencies
-        src_dir,derivatives = cap.scanDependencies(yml=False)
+        src_dir,derivatives = cap.scanDependencies(update=False)
         #write in all library and uses
         print(derivatives)
         libs = set()
@@ -76,31 +73,29 @@ class legoHDL:
             line = line.replace("template", cap.getName())
             if not addedCompDec and line.count("package") > 0:
                 addedCompDec = True
-                comp = cap.ports().replace("entity", "component")
+                comp = cap.ports(False)
                 line = line + "\n" + comp
                 pass
             vhd_pkg.write(line)
-            print(line,end='')
-        print()
         vhd_pkg.close()
         tmp_pkg.close()
         pass
 
-    #TO-DO: IMPLEMENT
     def install(self, cap, opt):
-        cache_path = self.hidden+"cache/"
-        lib_path = self.hidden+"lib/"+cap.getLib()+"/"
-        # grab the ID to search for the project
-        iden,rep = self.db.findPrj(cap.getLib(), cap.getName())
-        cap = caps.Capsule(rp=rep) #update info into a capsule obj
-        # clone the repo -> cache
-        if(iden != -1):
-            print(rep.name)
-        self.db.findProjectsLocal(cache_path, cached=True)
+        cache_path = apt.Apparatus.HIDDEN+"cache/"
+        lib_path = apt.Apparatus.HIDDEN+"lib/"+cap.getLib()+"/"
         #does the package already exist in the cache directory?
-        if(iden in self.db.getCachePrjs().keys()):
+        if(self.db.prjExists(cap.getTitle(), "cache")):
             print("Package is already installed.")
             return
+        elif(not self.db.prjExists(cap.getTitle(), "remote")):
+            pass
+     
+        # grab the ID to search for the project
+        iden,rep = self.db.findPrj(cap.getLib(), cap.getName())
+        cap = caps.Capsule(rp=rep,title=cap.getTitle()) #update info into a capsule obj
+        # clone the repo -> cache
+        
         #possibly make directory for cached project
         print("Installing... ",end='')
         cache_path = cache_path+cap.getLib()+"/"
@@ -117,24 +112,26 @@ class legoHDL:
         cap.install(cache_path, ver)
         print("success")
         print("library files path:")
-        #generate PKG VHD
-        self.genPKG(cap)
+
+        #generate PKG VHD    
+        self.genPKG(cap.getTitle())
         
         #link it all together through writing paths into "mapping.toml"
         cur_lines = list()
-        mapper = open(self.hidden+"map.toml", 'r')
+        mapper = open(apt.Apparatus.HIDDEN+"map.toml", 'r')
         cur_lines = mapper.readlines()
         mapper.close()
 
-        mapper = open(self.hidden+"map.toml", 'w')
+        mapper = open(apt.Apparatus.HIDDEN+"map.toml", 'w')
         inc_paths = list()
-        inc_paths.append("\'"+lib_path+"*.vhd"+"\',\n")
+        inc_paths.append("\'"+lib_path+cap.getName()+"_pkg.vhd"+"\',\n")
         inc_paths.append("\'"+cap.findPath(cap.getMeta("toplevel")).replace(cap.getMeta("toplevel"),"*.vhd")+"\',\n")
         inc = False
         found_lib = False
         if(len(cur_lines) <= 1):
             cur_lines.clear()
             mapper.write("[libraries]\n")
+
         for line in cur_lines:
             if(line.count(cap.getLib()+".files") > 0): #include into already established library section
                 inc = True
@@ -149,22 +146,37 @@ class legoHDL:
             mapper.write(line)
 
         if(len(cur_lines) == 0 or not found_lib):
-            if(line.count(cap.getLib()+".files") == 0): #create new library section
-                mapper.write(cap.getLib()+".files = [\n")
-                for p in inc_paths:
-                    mapper.write(p)
-                mapper.write("]\n")
+            #create new library section
+            mapper.write(cap.getLib()+".files = [\n")
+            for p in inc_paths:
+                mapper.write(p)
+            mapper.write("]\n")
 
         mapper.close()
         pass
 
-    #TO-DO: IMPLEMENT
     def uninstall(self, pkg, opt):
         #remove from cache
+        l,n = caps.Capsule.siftLibName(pkg)
+        if(self.db.prjExists(pkg, "cache")):
+            cache = self.db.getProjectsCache()
+            shutil.rmtree(cache[l][n])
+            #remove from lib
+            lib_path = cache[l][n].replace("cache","lib")
+            lib_path = lib_path[:len(lib_path)-1]+"_pkg.vhd"
+            os.remove(lib_path)
 
         #remove from 'mapping.toml'
-
-        #remove from lib
+        lines = list()
+        with open(apt.Apparatus.HIDDEN+"map.toml", 'r') as file:
+            lines = file.readlines()
+            file.close()
+        with open(apt.Apparatus.HIDDEN+"map.toml", 'w') as file:
+            for lin in lines:
+                if(lin.count(l) and (lin.count("/"+n+"/") or lin.count("/"+n+"_pkg"))):
+                    continue
+                file.write(lin)
+            file.close()
         pass
 
     def recurseScan(self, grp, d_list, top_mp):
@@ -174,10 +186,10 @@ class legoHDL:
         for d in d_list:
             l,n = caps.Capsule.siftLibName(d)
             n = n.replace("_pkg", "")
-            try:
-                with open(self.hidden+"cache/"+l+"/"+n+"/."+n+".yml", "r") as file:
+            if(os.path.isfile(apt.Apparatus.HIDDEN+"cache/"+l+"/"+n+"/."+n+".yml")):
+                with open(apt.Apparatus.HIDDEN+"cache/"+l+"/"+n+"/."+n+".yml", "r") as file:
                     tmp = yaml.load(file, Loader=yaml.FullLoader)
-            except:
+            else:
                 continue
             top_mp[l+'.'+n] = tmp['toplevel']
             grp,top_mp = self.recurseScan(grp, tmp['derives'], top_mp)
@@ -185,19 +197,15 @@ class legoHDL:
                 grp.addEdge(d, z)
         return grp,top_mp
 
-    def libExists(self, lib):
-        return os.path.isdir(self.hidden+"lib/"+lib)
-
     def export(self, cap, top=None, tb=None):
         print("Exporting...",end=' ')
-        build_dir = cap.getPath()+"/build/"
+        print(cap.getPath())
+        build_dir = cap.getPath()+"build/"
         #create a clean build folder
-        try:
-            os.mkdir(build_dir)
-        except:
+        if(os.path.isdir(build_dir)):
             shutil.rmtree(build_dir)
-            os.mkdir(build_dir)
-        
+        os.mkdir(build_dir)
+
         #add export option to override auto detection
         if(top == None):
             cap.autoDetectTop()
@@ -215,7 +223,7 @@ class legoHDL:
 
         g= graph.Graph()
         top_mp = dict()
-        top_mp[cap.getLib()+'.'+cap.getName()] = cap.getMeta('toplevel')
+        top_mp[cap.getTitle()] = cap.getMeta('toplevel')
         output = open(build_dir+"recipe", 'w')
         #mission: recursively search through every src VHD file for what else needs to be included
         src_dir,derivatives = cap.scanDependencies()
@@ -234,7 +242,7 @@ class legoHDL:
             n = n.replace("_pkg", "")
             #print(l,n)
             #library must exist in lib to be included in recipe.txt (avoids writing external libs like IEEE)
-            if(self.libExists(l)): #check lib exists
+            if(os.path.isdir(apt.Apparatus.HIDDEN+"lib/"+l)): #check lib exists
                 if not l in library.keys():
                     library[l] = list() 
                 library[l].append(n)
@@ -243,10 +251,10 @@ class legoHDL:
         for lib in library.keys():
             for pkg in library[lib]:
                 key = lib+'.'+pkg
-                root_dir = self.hidden+"cache/"+lib+"/"+pkg+"/"
+                root_dir = apt.Apparatus.HIDDEN+"cache/"+lib+"/"+pkg+"/"
                 src_dir = glob.glob(root_dir+"/**/"+top_mp[key], recursive=True)
                 output.write("LIB "+lib+" "+src_dir[0].replace(top_mp[key], "*.vhd")+"\n")
-                output.write("LIB "+lib+" "+self.hidden+"lib/"+lib+"/"+pkg+"_pkg.vhd\n")
+                output.write("LIB "+lib+" "+apt.Apparatus.HIDDEN+"lib/"+lib+"/"+pkg+"_pkg.vhd\n")
 
         #write current src dir where all src files are as "work" lib
         output.write("SRC "+cap.findPath(top).replace(top, "*.vhd")+"\n")
@@ -268,16 +276,15 @@ class legoHDL:
             print('ERROR- Package \''+cap.getName()+'\' does not exist in remote')
             return
 
-        if iden in self.db.getCurPrjs().keys(): #just give it an update!
+        if(self.db.prjExists(cap.getTitle(), "local")):
             print("Project already exists in local workspace- pulling from remote...",end=' ')
-            cap.pull() 
-        else: #oh man, go grab the whole thing!
+            cap.pull()
+        else:
             print("Cloning from remote...",end=' ')
             cap.clone()
-        
     
         try: #symbolically link the downloaded project folder to the cache
-            shutil.rmtree(self.hidden+"cache/"+cap.getLib()+"/"+cap.getName()+"/")
+            shutil.rmtree(apt.Apparatus.HIDDEN+"cache/"+cap.getLib()+"/"+cap.getName()+"/")
         except:
             pass
         #generate PKG VHD
@@ -306,11 +313,11 @@ class legoHDL:
         cap.autoDetectTop()
         cap.release(ver, options)
 
-        try: shutil.rmtree(self.hidden+"cache/"+cap.getLib()+"/"+cap.getName())
+        try: shutil.rmtree(apt.Apparatus.HIDDEN+"cache/"+cap.getLib()+"/"+cap.getName())
         except: pass
         #clone new project's progress into cache
         tmp = caps.Capsule(cap.getLib()+'.'+cap.getName())
-        tmp.install(self.hidden+"cache/"+cap.getLib()+"/", "v"+cap.getVersion(), src_url=cap.getPath())
+        tmp.install(apt.Apparatus.HIDDEN+"cache/"+cap.getLib()+"/", "v"+cap.getVersion(), src_url=cap.getPath())
         self.genPKG(tmp)
 
         if caps.Capsule.linkedRemote():
@@ -329,7 +336,7 @@ class legoHDL:
             self.db.encrypt(choice, options[0])
             return
 
-        if(not options[0] in self.settings.keys()):
+        if(not options[0] in apt.Apparatus.SETTINGS.keys()):
             print("ERROR- Invalid setting")
             return
 
@@ -344,11 +351,10 @@ class legoHDL:
         if(options[0] == 'local'):
             os.makedirs(choice, exist_ok=True)
 
-        self.settings[options[0]] = choice
+        apt.Apparatus.SETTINGS[options[0]] = choice
 
-        with open(self.pkgmngPath+"settings.yml", "w") as file:
-            yaml.dump(self.settings, file)
-            pass
+        apt.Apparatus.save()
+        pass
     
     #TO-DO: implement
     def convert(self, package):
@@ -370,7 +376,7 @@ class legoHDL:
             print('No module '+cap.getName()+' exists locally')
             return
         
-        if(self.remote == None and force):
+        if(apt.Apparatus.SETTINGS['remote'] == None and force):
             print('\
                 WARNING- No remote code base is configured, if this module is deleted it may be unrecoverable.\n \
                 DELETE '+cap.getName()+'? [y/n]\
@@ -400,23 +406,8 @@ class legoHDL:
         pkgPath = os.getcwd()
         lastSlash = pkgPath.rfind('/') #determine project's name to know the YAML to open
         pkgCWD = pkgPath[lastSlash+1:]
-        self.capsuleCWD = caps.Capsule(pkgCWD)
-        #is there a .yaml here? if so, grab the id and then load the project from the repo
-        if(os.path.isfile(pkgPath+'/.'+pkgCWD+".yml")):
-            with open(pkgPath+'/.'+pkgCWD+".yml", 'r') as f:
-                tmp = yaml.load(f, Loader=yaml.FullLoader)
-                f.close()
-                pass
-            if tmp['id'] in self.db.getCurPrjs().keys():
-                self.capsuleCWD = caps.Capsule(rp=self.db.getCurPrjs()[tmp['id']])
-            pass
 
-        if(not self.capsuleCWD.isValid()):
-            #print("NOT A CAPSULE DIRECTORY")
-            pass
-        else:
-            #print("VALID CAPSULE DIRECTORY")
-            pass
+        self.capsuleCWD = caps.Capsule(name=pkgCWD,path=pkgPath+"/")
         
         command = package = description = ""
         options = []
@@ -440,10 +431,11 @@ class legoHDL:
                 break
 
         package = package.replace("-", "_")
-        self.capsulePKG = caps.Capsule(package)
+        self.capsulePKG = caps.Capsule(title=package)
         
         #branching through possible commands
         if(command == "install"):
+            print(self.capsulePKG.getTitle())
             self.install(self.capsulePKG, options) #TO-DO
             pass
         elif(command == "uninstall"):
@@ -495,7 +487,7 @@ class legoHDL:
             pass
         elif(command == "init"):
             self.convert(package)
-        elif(command == "export"): #a visual aide to help a developer see what package's are at the ready to use
+        elif(command == "export" and self.capsuleCWD.isValid()): #a visual aide to help a developer see what package's are at the ready to use
             #'' and list() are default to pkg and options
             mod = package
             tb = None
@@ -511,13 +503,20 @@ class legoHDL:
         elif(command == "show" and self.capsulePKG.isValid()):
             self.capsulePKG.show()
             pass
-        elif(command == "ports" and self.capsulePKG.isValid()):
+        elif(command == "ports"):
             mapp = False
             if(len(options) and 'map' in options):
                 mapp = True
-            print(self.capsulePKG.ports(mapp))
-        elif(command == "template" and self.settings['editor'] != None):
-            os.system(self.settings['editor']+" "+self.pkgmngPath+"/template")
+            if(self.capsulePKG.isValid()):
+                print(self.capsulePKG.ports(mapp))
+            else:
+                l,n = caps.Capsule.siftLibName(package)
+                cache = self.db.getProjectsCache()
+                if(l in cache.keys() and n in cache[l].keys()):
+                    print(caps.Capsule(name=n,path=cache[l][n]).ports(mapp))
+                    pass
+        elif(command == "template" and apt.Apparatus.SETTINGS['editor'] != None):
+            os.system(apt.Apparatus.SETTINGS['editor']+" "+apt.Apparatus.PKGMNG_PATH+"/template")
             pass
         elif(command == "config"):
             self.setSetting(options, package)
@@ -586,7 +585,9 @@ class legoHDL:
 
 def main():
     #print('\n---legoHDL package manager---\n')
+    apt.Apparatus.load()
     legoHDL()
+
 
 
 if __name__ == "__main__":
